requirements.txt
torch>=2.1
transformers>=4.38
accelerate>=0.25
pillow
openai>=1.12
google-generativeai>=0.3.0
Pillow>=10.0.0
